RavenDB-19371 - rename ClosedDueNoDocsLeft to NoDocsLeft
