+++ src/Raven.Client/Exceptions/Documents/Subscriptions/SubscriptionClosedException.cs
@@ -11,7 +11,7 @@ namespace Raven.Client.Exceptions.Documents.Subscriptions
     public class SubscriptionClosedException : SubscriptionException
     {
         internal bool CanReconnect { get; set; }
-        internal bool ClosedDueNoDocsLeft { get; set; }
+        internal bool NoDocsLeft { get; set; }
 
         public SubscriptionClosedException(string message) : base(message)
         {
@@ -22,10 +22,10 @@ namespace Raven.Client.Exceptions.Documents.Subscriptions
             CanReconnect = canReconnect;
         }
 
-        public SubscriptionClosedException(string message, bool canReconnect, bool closedDueNoDocsLeft) : base(message)
+        public SubscriptionClosedException(string message, bool canReconnect, bool noDocsLeft) : base(message)
         {
             CanReconnect = canReconnect;
-            ClosedDueNoDocsLeft = closedDueNoDocsLeft;
+            NoDocsLeft = noDocsLeft;
         }
 
         public SubscriptionClosedException(string message, Exception inner) : base(message, inner)
@@ -37,10 +37,10 @@ namespace Raven.Client.Exceptions.Documents.Subscriptions
             CanReconnect = canReconnect;
         }
 
-        public SubscriptionClosedException(string message, bool canReconnect, bool closedDueNoDocsLeft, Exception inner) : base(message, inner)
+        public SubscriptionClosedException(string message, bool canReconnect, bool noDocsLeft, Exception inner) : base(message, inner)
         {
             CanReconnect = canReconnect;
-            ClosedDueNoDocsLeft = closedDueNoDocsLeft;
+            NoDocsLeft = noDocsLeft;
         }
     }
 }
