+++ src/Raven.Server/Documents/TcpHandlers/SubscriptionConnection.cs
@@ -11,10 +11,8 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using Esprima;
-using Raven.Client;
 using Raven.Client.Documents.Subscriptions;
 using Raven.Client.Exceptions;
-using Raven.Client.Exceptions.Documents.Subscriptions;
 using Raven.Client.ServerWide;
 using Raven.Server.Documents.Queries;
 using Raven.Server.Documents.Queries.AST;
